Enforce path (de)normalisation (#28)
Internally, Isabelle aggressively use canonical names and not absolute
name -- except when opening a file in the LSP server. Now, on the
Isabelle side, we always use the canonical names.

On the emacs, we cannot (at least not without highly invasive
changes), so we only normalize paths back and forth when sending paths
to the Isabelle LSP server. This is done by
`lsp-isar-file-name-unfollow-links` and must be customized to work.

The main drawback is that sending message is a bit slower now, since
the normalization needs to access the file system.